Reorganize, add setup.py so it can be packaged
I added a setup.py and setup.cfg so this project can be packaged into an
sdist/egg/wheel. I also moved stuff around to make it easier to package
and I converted README.md into a README.rst, because PyPI is currently
much more friendly to reST than to Markdown.

With this you should be able to:

    python setup.py install
        - or -
    python setup.py develop
        - or -
    pip install -e .